+++ lucid/lucid/renderer/vector/vector4d.h
@@ -1,9 +1,9 @@
 #pragma once
 
 struct vec4_t {
-    float x, y, z, rhw;
+    double x, y, z, rhw;
 
     vec4_t() {}
-    vec4_t(float _x, float _y, float _z, float _rhw)
+    vec4_t(double _x, double _y, double _z, double _rhw)
         : x(_x), y(_y), z(_z), rhw(_rhw) {}
 };
